Add devstack settings file
