+++ analytics_dashboard/settings/production.py
@@ -1,14 +1,7 @@
 """Production settings and globals."""
 
-from os import environ
-
-# Normally you should not import ANYTHING from Django directly
-# into your settings, but ImproperlyConfigured is an exception.
-from django.core.exceptions import ImproperlyConfigured
-
-import yaml
-
 from analytics_dashboard.settings.base import *
+from analytics_dashboard.settings.yaml_config import *
 from analytics_dashboard.settings.logger import get_logger_config
 
 # Enable offline compression of CSS/JS
@@ -26,25 +19,10 @@ COMPRESS_CSS_FILTERS += [
 LOGGING = get_logger_config()
 
 
-def get_env_setting(setting):
-    """ Get the environment setting or return exception """
-    try:
-        return environ[setting]
-    except KeyError:
-        error_msg = "Set the %s env variable" % setting
-        raise ImproperlyConfigured(error_msg)
-
 # ######### HOST CONFIGURATION
 # See: https://docs.djangoproject.com/en/1.5/releases/1.5/#allowed-hosts-required-in-production
 ALLOWED_HOSTS = ['*']
 ########## END HOST CONFIGURATION
-
-CONFIG_FILE = get_env_setting('ANALYTICS_DASHBOARD_CFG')
-
-with open(CONFIG_FILE) as f:
-    config_from_yaml = yaml.load(f)
-
-vars().update(config_from_yaml)
 
 DB_OVERRIDES = dict(
     PASSWORD=environ.get('DB_MIGRATION_PASS', DATABASES['default']['PASSWORD']),
